requirements.txt
absl-py==1.3.0
accelerate==0.15.0
aiohttp==3.8.3
aiosignal==1.2.0
albumentations==1.3.0
alembic==1.8.1
altair==4.2.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
autopage==0.5.1
av==10.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
bitarray==2.6.2
black==22.10.0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
cliff==4.0.0
cmaes==0.8.2
cmake==3.25.2
cmd2==2.4.2
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.3
decorator==4.4.2
decord==0.6.0
defusedxml==0.7.1
diffusers==0.11.1
distlib==0.3.6
docker-pycreds==0.4.0
einops==0.6.0
ema-pytorch==0.2.2
entrypoints==0.4
exceptiongroup==1.0.0
executing==1.2.0
face-alignment==1.3.5
facenet-pytorch==2.5.2
fastapi==0.89.1
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.8.0
flake8==5.0.4
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.1
fsspec==2022.10.0
future==0.18.2
gdown==4.6.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.0
google-auth-oauthlib==0.4.6
gradio==3.16.1
greenlet==2.0.0
grpcio==1.50.0
h11==0.14.0
h5py==3.7.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.11.1
hydra-colorlog==1.2.0
hydra-core==1.2.0
hydra-optuna-sweeper==1.2.0
identify==2.5.8
idna==3.4
imageio==2.22.3
imageio-ffmpeg==0.4.7
importlib-metadata==5.0.0
iniconfig==1.1.1
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
kiwisolver==1.4.4
kornia==0.6.9
lightning-utilities==0.8.0
linkify-it-py==1.0.3
lit==15.0.7
llvmlite==0.39.1
lxml==4.9.2
Mako==1.2.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.4
more-itertools==9.0.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.2
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.8.8
ninja==1.11.1
nodeenv==1.7.0
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.2.3
openai-whisper==20230314
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.68
optuna==2.10.1
orjson==3.8.5
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pathtools==0.1.2
pbr==5.11.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.2
pluggy==1.0.0
portalocker==2.7.0
pre-commit==2.20.0
prettytable==3.5.0
proglog==0.1.10
progressbar2==4.2.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.4
pyDeprecate==0.3.2
pydub==0.25.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pyperclip==1.8.2
pyre-extensions==0.0.23
pyrootutils==1.0.4
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
python-multipart==0.0.5
python-utils==3.4.5
pytorch-lightning==2.0.0
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==12.6.0
rotary-embedding-torch==0.1.5
rsa==4.9
sacrebleu==2.3.1
scikit-image==0.19.3
scikit-learn==1.2.0
scikit-video==1.1.11
scipy==1.9.3
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.10.1
setproctitle==1.3.2
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
stack-data==0.6.0
starlette==0.22.0
stevedore==4.1.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tiktoken==0.3.1
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
torch-fidelity==0.3.0
torchaudio==2.0.1
torchmetrics==0.11.4
torchvision==0.15.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
triton==2.0.0
typing-inspect==0.8.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
uri-template==1.2.0
urllib3==1.26.12
uvicorn==0.20.0
virtualenv==20.16.6
wandb==0.13.4
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
Werkzeug==2.2.2
widgetsnbextension==4.0.5
x-transformers==1.5.1
yarl==1.8.1
zipp==3.10.0
git+https://github.com/DinoMan/dino-tk.git@master#egg=dtk
clint
soundfile
audiomentations
# git+https://github.com/antonibigata/video-audio-matching.git#egg=video_audio_matching